allowed distance to be floating point as well as hours slept on edit page and reflected this change on charts(parse as float instead of int)

+++ Database_Docs/Dump20170426.sql
@@ -289,7 +289,7 @@ CREATE TABLE `Workouts` (
   `wid` int(11) NOT NULL AUTO_INCREMENT,
   `uid` int(11) NOT NULL,
   `date` datetime NOT NULL,
-  `sleep` int(11) DEFAULT NULL,
+  `sleep` float DEFAULT NULL,
   `health` varchar(45) DEFAULT NULL,
   `injury` varchar(45) DEFAULT NULL,
   `percent_health` varchar(45) DEFAULT NULL,
